Stuff actually commits now, but a bunch of repositories are busted

+++ src/main/java/p2/backend/Beans/Food.java
@@ -28,7 +28,7 @@ public class Food {
 
     @Id
     @GeneratedValue// @Column(name = "foodId")
-    private Integer foodId;
+    private Long foodId;
 
    // @Column(name = "foodName")
     private String foodName;
